refactor: page, layout 스타일 파일 분리
