+++ src/app/daily/layout.styled.ts
@@ -1,0 +1,12 @@
+'use client';
+
+import styled from '@emotion/styled';
+
+export const DailyWrapper = styled.main`
+  width: 100%;
+  height: 100%;
+`;
+
+export const Spacer = styled.div`
+  height: 20rem;
+`;
