+++ src/app/daily/page.styled.ts
@@ -2,17 +2,6 @@
 
 import styled from '@emotion/styled';
 
-// layout
-export const DailyWrapper = styled.main`
-  width: 100%;
-  height: 100%;
-`;
-
-export const Spacer = styled.div`
-  height: 20rem;
-`;
-
-// page
 export const DailyPageWrapper = styled.div`
   display: flex;
   flex-direction: column;
